Global Styles: Extract shared ColorGradientDropdownItem
Pull the ColorPanelDropdown component and its supporting
LabeledColorIndicators/ColorPanelTab/popoverProps helpers out of
color-panel.js into a standalone module so Background, Typography, and
Color panels can all render color/gradient ToolsPanelItems consistently
without duplicating the dropdown/tabs/reset boilerplate.

Pure refactor. No UI change.
